pyproject.toml
[project]
name = "monobank-mcp"
version = "0.1.0"
description = "MonoBank MCP Server"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Nazar Klymok", email = "my@email.com"},
]
dependencies = [
    "fastmcp",
    "pydantic",
    "httpx",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.4"

[tool.uv.sources]
"fastmcp" = { url = "https://pypi.fastmcp.com/simple" }
